+++ management/info.sh
@@ -18,4 +18,7 @@ Admin Username:   admin
 Admin Password:   $ADMIN_PASSWORD
 Management IP:    $CPMAN_IP
 
+Commands:
+    ssh admin@$CPMAN_IP
+    ssh-copy-id admin@$CPMAN_IP   # to setup key-based auth (assuming you have a key in ~/.ssh/id_rsa.pub)
 EOF