+++ singlegw/info.sh
@@ -7,6 +7,7 @@ NAME=$(terraform output -raw name)
 ADMIN_PASSWORD=$(terraform output -raw admin_password)
 
 GATEWAY_IP=$(az vm show -d --resource-group "$RG" --name "$NAME" --query "publicIps" -o tsv)
+SIC_KEY=$(terraform output -raw sic_key)
 
 cat <<EOF
 
@@ -17,9 +18,10 @@ Name:             $NAME
 Admin Username:   admin
 Admin Password:   $ADMIN_PASSWORD
 Gateway IP:       $GATEWAY_IP
+SIC key:          $SIC_KEY
 
 Commands:
-    ssh admin@"$GATEWAY_IP"
-    ssh-copy-id admin@"$GATEWAY_IP"   # to setup key-based auth (assuming you have a key in ~/.ssh/id_rsa.pub)
+    ssh admin@$GATEWAY_IP
+    ssh-copy-id admin@$GATEWAY_IP   # to setup key-based auth (assuming you have a key in ~/.ssh/id_rsa.pub)
 
 EOF